+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- On Wayland, added a `get_wayland_display` function to `EventsLoopExt`.
 - On Windows, fix `CursorMoved(0, 0)` getting dispatched on window focus.
 - On macOS, fix command key event left and right reverse.
 - On FreeBSD, NetBSD, and OpenBSD, fix build of X11 backend.